offset working

+++ gpuowl.cl
@@ -394,7 +394,7 @@ KERNEL(256) fft2K_1K(CONST double2 *in, global double2 *out, SMALL_CONST double2
 }
 
 // conjugates input
-KERNEL(256) carryA(const uint baseBits, const uint offsetWord, const long offsetVal,
+KERNEL(256) carryA(const uint baseBits, const uint offsetWord, const double offsetVal,
                    CONST double2 *in, CONST double2 *A, global int2 *out,
                    global long *carryOut, global uint *globalMaxErr) {
   uint g  = get_group_id(0);
@@ -410,7 +410,7 @@ KERNEL(256) carryA(const uint baseBits, const uint offsetWord, const long offset
 
   for (int i = 0; i < 8; ++i) {    
     uint p = me + i * 1024;
-    if (g / 4 * 8 + g % 4 * 256 * 2048 + i + me * 2048 == offsetWord) { carry += offsetVal; }
+    if (g / 4 * 8 + g % 4 * 256 * 2048 + i + me * 2048 == offsetWord) { carry += (long) offsetVal; }
     out[p] = car0(&carry, conjugate(in[p]), A[p], &maxErr, baseBits);
   }
